feat: add support for custom indexing logic in Select
This commit adds support for custom indexing logic to the `Select` component. This allows developers to control how the search results are ranked, which can be useful for scenarios where the default ranking algorithm is not sufficient.

The `computeIndexingScore` function can be passed to the `SelectItem` widget. This function will be called whenever the search query changes, and it should return an integer value representing the relevance of the item to the current query. The higher the score, the more relevant the item is considered to be.