Fix: Allow Select to accept String type
The commit changed the type of the `selectedValue` from `int?` to `String?`. This change allows the `Select` widget to accept values of type `String`.

It also adds a `showUnrelatedValues` property to the `Select` widget. This property can be used to control whether or not the widget displays values that are not related to the search query.