feat: Add `onChanged` callback to `Select`
This commit introduces a new `onChanged` callback to the `Select` component. This callback is fired when the selected value changes, allowing for more flexible handling of state updates.

The `onChanged` callback is optional. If it is not provided, the `Select` component will still allow the user to select a value, but no state updates will occur.

This change also removes the `onChanged` callback from the `SelectPopup` component, as this callback is now handled by the `Select` component itself.

This change improves the flexibility and usability of the `Select` component.