feat(menu): Add support for submenus
This commit introduces a new `MenuPopup` widget for creating submenus. The `MenuPopup` widget is used in conjunction with the `Popover` widget to display a submenu when a menu item is clicked.

This commit also adds a new `dataBuilder` property to the `MenuGroup` widget, which allows the user to specify how the data for each menu item should be created.